refactor: convert service layer from interface pattern to concrete class
- Removed OrderService and UserService interfaces, converted to concrete @Service classes
- Deleted OrderServiceImpl and UserServiceImpl
- Renamed test files and updated them to test concrete classes
- Added verifyNoMoreInteractions() calls in service unit tests
- Updated AGENTS.md documentation
