--- order-api/src/main/java/com/ivanfranchin/orderapi/order/OrderServiceImpl.java
@@ -1,44 +1,0 @@
-package com.ivanfranchin.orderapi.order;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@RequiredArgsConstructor
-@Service
-public class OrderServiceImpl implements OrderService {
-
-    private final OrderRepository orderRepository;
-
-    @Override
-    public List<Order> getOrders() {
-        return orderRepository.findAllByOrderByCreatedAtDesc();
-    }
-
-    @Override
-    public long countOrders() {
-        return orderRepository.count();
-    }
-
-    @Override
-    public List<Order> getOrdersContainingText(String text) {
-        return orderRepository.findByIdContainingOrDescriptionContainingIgnoreCaseOrderByCreatedAt(text, text);
-    }
-
-    @Override
-    public Order validateAndGetOrder(String id) {
-        return orderRepository.findById(id)
-                .orElseThrow(() -> new OrderNotFoundException("Order with id %s not found".formatted(id)));
-    }
-
-    @Override
-    public Order saveOrder(Order order) {
-        return orderRepository.save(order);
-    }
-
-    @Override
-    public void deleteOrder(Order order) {
-        orderRepository.delete(order);
-    }
-}
