--- order-api/src/main/java/com/ivanfranchin/orderapi/user/UserServiceImpl.java
@@ -1,61 +1,0 @@
-package com.ivanfranchin.orderapi.user;
-
-import com.ivanfranchin.orderapi.security.Role;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.Optional;
-
-@RequiredArgsConstructor
-@Service
-public class UserServiceImpl implements UserService {
-
-    private final UserRepository userRepository;
-
-    @Override
-    public List<User> getUsers() {
-        return userRepository.findAllByOrderByUsernameAsc();
-    }
-
-    @Override
-    public long countUsers() {
-        return userRepository.count();
-    }
-
-    @Override
-    public long countAdmins() {
-        return userRepository.countByRole(Role.ADMIN);
-    }
-
-    @Override
-    public Optional<User> getUserByUsername(String username) {
-        return userRepository.findByUsername(username);
-    }
-
-    @Override
-    public boolean hasUserWithUsername(String username) {
-        return userRepository.existsByUsername(username);
-    }
-
-    @Override
-    public boolean hasUserWithEmail(String email) {
-        return userRepository.existsByEmail(email);
-    }
-
-    @Override
-    public User validateAndGetUserByUsername(String username) {
-        return getUserByUsername(username)
-                .orElseThrow(() -> new UserNotFoundException("User with username %s not found".formatted(username)));
-    }
-
-    @Override
-    public User saveUser(User user) {
-        return userRepository.save(user);
-    }
-
-    @Override
-    public void deleteUser(User user) {
-        userRepository.delete(user);
-    }
-}
